Separate Performance Tests into reasonable pieces
1. introduce helper for grabbing config
2. introduce a second performance test case
3. make the console app entrance cleaner
