+++ tools/Performance/DynamoPerformanceTests/DynamoPerformanceTests.csproj
@@ -45,7 +45,7 @@
     <Reference Include="DynamoCoreTests">
       <HintPath>..\..\..\bin\AnyCPU\$(Configuration)\DynamoCoreTests.dll</HintPath>
     </Reference>
-        <Reference Include="DynamoCoreWpfTests">
+    <Reference Include="DynamoCoreWpfTests">
       <HintPath>..\..\..\bin\AnyCPU\$(Configuration)\DynamoCoreWpfTests.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.CodeAnalysis, Version=2.10.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
@@ -220,8 +220,11 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DynamoModelPerformanceTestBase.cs" />
     <Compile Include="GraphNameColumn.cs" />
-    <Compile Include="PerformanceTestFramework.cs" />
+    <Compile Include="DynamoViewModelPerformanceTestBase.cs" />
+    <Compile Include="PerformanceTestConsoleApp.cs" />
+    <Compile Include="PerformanceTestHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="..\..\..\src\AssemblySharedInfoGenerator\AssemblySharedInfo.cs">
       <Link>Properties\AssemblySharedInfo.cs</Link>
